move config to own module
